+++ src/pages/calendar/CalendarPage.js
@@ -3,29 +3,25 @@ import styled from "styled-components";
 import { Link } from "react-router-dom";
 
 import Calendar from "../../components/calendar/Calendar";
-import Schedule from "../../components/calendar/Schedule";
+import ScheduleList from "../../components/calendar/ScheduleList";
 
 // icon, bg-color
-import { FaRegAddressBook } from "react-icons/fa";
-import { AiOutlineCalendar } from "react-icons/ai";
-import { RiMapPinLine, RiPlayList2Fill } from "react-icons/ri";
-import { HiOutlineMail } from "react-icons/hi";
-import { BiUserCircle } from "react-icons/bi";
+
 
 import color from "./../../assets/color.png";
+import Header from '../../common/Header';
 
-
-
+// CSS
 const Test = styled.div`
   width: 100vw;
   height: 100vh;
   background: #f5f3fe;
 `;
+
 const MainDiv = styled.div`
   position: absolute;
   width: 90vw;
   height: 85vh;
-
   margin-top: 78px;
   margin-left: 69px;
   margin-right: 69px;
@@ -34,8 +30,9 @@ const MainDiv = styled.div`
   border-radius: 20px;
   display: flex;
   flex-direction: row;
-  z-index: 2;
+  z-index: 1;
 `;
+
 const BackColor = styled.img`
   position: absolute;
   width: 548px;
@@ -43,30 +40,30 @@ const BackColor = styled.img`
   margin-left: 100px;
   margin-top: 100px;
   background: #f8f8f8;
-  z-index: 1;
-`;
-
-const Header = styled.div`
-  background: #f8f8f8;
-  width: 10%;
-  display: flex;
-  flex-direction: column;
-  border-radius: 20px;
-  z-index: 4;
+  z-index: 2;
 `;
 
 const Mid = styled.div`
   background: #f8f8f8;
   width: 30%;
   border-radius: 20px;
-  /* overflow: scroll; */
 `;
+
+const ScheduleBox = styled.div`
+  width: 100%;
+  height: 503px;
+  border-radius: 20px;
+  position: relative;
+  z-index: 3;
+`;
+
 const Title = styled.div`
   margin-top: 50px;
   font-size: 20px;
   margin-top: 20px;
   margin-bottom: 5px;
 `;
+
 const SubTitle = styled.div`
   margin-bottom: 38px;
 `;
@@ -84,32 +81,13 @@ const CalendarPage = () => {
     <Test>
       <MainDiv className="MainDiv">
         <BackColor src={color} style={{ opacity: 0.2 }} />
-        <Header>
-          <Link to="/">
-            <AiOutlineCalendar className="true" />
-          </Link>
-          <Link to="/contact">
-            <FaRegAddressBook className="false" />
-          </Link>
-          <Link to="/reservation">
-            <RiMapPinLine className="false" />
-          </Link>
-          <Link to="/approval">
-            <HiOutlineMail className="false" />
-          </Link>
-          <Link to="/minutes">
-            <RiPlayList2Fill className="false" />
-          </Link>
-          <Link to="/profile">
-            <BiUserCircle className="false" />
-          </Link>
-        </Header>
+        <Header />
         <Mid>
           <Title>오늘의 일정</Title>
           <SubTitle>It's today's schedule.</SubTitle>
-          <Schedule />
-          <Schedule />
-          <Schedule />
+          <ScheduleBox>
+            <ScheduleList />
+          </ScheduleBox>
         </Mid>
         <Last>
           <Calendar />
